+++ src/main/java/com/pool/repository/PlanRepository.java
@@ -11,7 +11,7 @@ import com.pool.domine.Plan;
 import com.pool.model.PlanModel;
 
 @Repository
-public interface PlanRepository extends JpaRepository<Plan, Long> {
+public interface PlanRepository extends JpaRepository<Plan, Integer> {
 	@Query(name="plantitle_description_query",nativeQuery = true)
 	public List<PlanModel> getCustomPlanData(); 
 }
